core/requirements.txt
pydantic
python-dotenv
fastapi[standard]
uvicorn
requests
mangum